move things to private
