+++ lib/rubocop/packs.rb
@@ -21,40 +21,6 @@ module RuboCop
 
     private_constant(:CONFIG_DEFAULT, :PROJECT_ROOT)
 
-    class Offense < T::Struct
-      extend T::Sig
-
-      const :cop_name, String
-      const :filepath, String
-
-      sig { returns(ParsePackwerk::Package) }
-      def pack
-        ParsePackwerk.package_from_path(filepath)
-      end
-    end
-
-    sig { params(paths: T::Array[String], cop_names: T::Array[String]).returns(T::Array[Offense]) }
-    def self.offenses_for(paths:, cop_names:)
-      path_arguments = paths.join(' ')
-      cop_arguments = cop_names.join(',')
-      # I think we can potentially use `RuboCop::CLI.new(args)` for this to avoid shelling out and starting another process that needs to reload the bundle
-      args = [path_arguments, "--only=#{cop_arguments}", '--format=json']
-      puts "Executing: bundle exec rubocop #{args.join(' ')}"
-      json = JSON.parse(Private.execute_rubocop(args))
-      offenses = T.let([], T::Array[Offense])
-      json['files'].each do |file_hash|
-        filepath = file_hash['path']
-        file_hash['offenses'].each do |offense_hash|
-          offenses << Offense.new(
-            cop_name: offense_hash['cop_name'],
-            filepath: filepath
-          )
-        end
-      end
-
-      offenses
-    end
-
     #
     # Ideally, this is API that is available to us via `rubocop` itself.
     # That is: the ability to preserve the location of `.rubocop_todo.yml` files and associate
@@ -63,7 +29,7 @@ module RuboCop
     sig { params(packs: T::Array[ParsePackwerk::Package], files: T::Array[String]).void }
     def self.regenerate_todo(packs: [], files: [])
       paths = packs.empty? ? files : packs.map(&:name).reject { |name| name == ParsePackwerk::ROOT_PACKAGE_NAME }
-      offenses = offenses_for(
+      offenses = Private.offenses_for(
         paths: paths,
         cop_names: config.permitted_pack_level_cops
       )
@@ -81,23 +47,6 @@ module RuboCop
         else
           rubocop_todo = {}
         end
-
-        offenses_for_pack.each do |offense|
-          rubocop_todo[offense.cop_name] ||= { 'Exclude' => [] }
-          rubocop_todo[offense.cop_name]['Exclude'] << offense.filepath
-        end
-
-        next if rubocop_todo.empty?
-
-        rubocop_todo_yml.write(YAML.dump(rubocop_todo))
-      end
-    end
-
-    sig { params(offenses: T::Array[Offense]).void }
-    def self.add_offenses(offenses)
-      offenses.group_by(&:pack).each do |pack, offenses_for_pack|
-        rubocop_todo_yml = pack.directory.join(PACK_LEVEL_RUBOCOP_TODO_YML)
-        rubocop_todo = rubocop_todo_yml.exist? ? YAML.load_file(rubocop_todo_yml) : {}
 
         offenses_for_pack.each do |offense|
           rubocop_todo[offense.cop_name] ||= { 'Exclude' => [] }
